+++ R/bootstrap.R
@@ -208,7 +208,7 @@ tableAUC <- function(SE_scored, annotationColName, signatureColNames,
 #'
 #' @examples
 #' # Run signature profiling
-#'  choose_sigs <- TBsignatures[c(1, 2)]
+#'  choose_sigs <- TBsignatures[c("Zak_RISK_16", "Zhao_NANO_6")]
 #'  prof_indian <- runTBsigProfiler(TB_indian[seq_len(25), ],
 #'                                  useAssay = "logcounts",
 #'                                  algorithm = "ssGSEA",